// src/base-ed/WechatUser/Schema.ts
export type OpSchema = {
    id: PrimaryKey;
    $$createAt$$: Datetime;
    $$updateAt$$: Datetime;
    $$removeAt$$?: Datetime | null;
    origin: 'mp' | 'public';
    openId?: String<32> | null;
    unionId?: String<32> | null;
    accessToken: String<32>;
    sessionKey?: String<64> | null;
    subscribed?: Boolean | null;
    subscribedAt?: Datetime | null;
    unsubscribedAt?: Datetime | null;
    userId?: ForeignKey<"user"> | null;
    applicationId: ForeignKey<"application">;
};